Extract metadata diff
