+++ kahuna/public/js/edits/service.js
@@ -5,6 +5,7 @@ import {editsApi} from '../services/api/edits-api';
 import {mediaApi} from '../services/api/media-api';
 import { overwrite, prepend, append } from '../util/constants/editOptions';
 import { trackAll } from '../util/batch-tracking';
+import { getMetadataDiff } from './metadataDiff';
 
 export var service = angular.module('kahuna.edits.service', [
     editsApi.name,
@@ -232,33 +233,7 @@ service.factory('editsService',
             .then(() => true, () => false);
     }
 
-    function getMetadataDiff (image, metadata) {
-        var diff = {};
 
-        // jscs has a maximumLineLength of 100 characters, hence the line break
-        var keys = new Set(Object.keys(metadata).concat(
-            Object.keys(image.data.originalMetadata)));
-
-        // Keywords, peopleInImage and subjects are arrays, the comparison below only works with string comparison.
-        // For simplicity, ignore them as we're not updating these fields at the moment.
-        keys.delete('keywords');
-        keys.delete('subjects');
-        keys.delete('peopleInImage');
-
-        keys.forEach((key) => {
-            if (metadata[key] && angular.isUndefined(image.data.originalMetadata[key])) {
-                diff[key] = metadata[key];
-            } else if (metadata[key] !== image.data.originalMetadata[key] &&
-                angular.isDefined(image.data.originalMetadata[key])) {
-                // if the user has provided an override of '' (e.g. they want remove the title),
-                // angular sets the value in the object to undefined.
-                // We need to use an empty string in the PUT request to obey user input.
-                diff[key] = metadata[key] || '';
-            }
-        });
-
-        return diff;
-    }
 
     function updateMetadataField (image, field, value) {
         var metadata = image.data.metadata;
